BAnk Page Update

+++ database/migrations/2024_04_24_063523_create_employee_salaries_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->decimal('debit', 12, 2)->nullable()->comment('Submit Salary');
             $table->decimal('creadit', 12, 2)->nullable()->comment('Employee Salary on Employee Salary Table');
             $table->decimal('balance', 12, 2)->comment('creadit - debit');
-            $table->integer('payment_method');
+            $table->bigInteger('payment_method');
             $table->text('note')->nullable();
             $table->timestamps();
         });
